requirements.txt
allure_python_commons==2.13.5
Faker==25.5.0
pytest==8.1.1
Requests==2.32.3
selenium==4.21.0
